feat(26/sl-ikzl): satsuma-core — main extraction functions
Add satsuma-core/src/extract.ts with all extraction functions:
extractSchemas, extractMappings, extractFragments, extractMetrics,
extractTransforms, extractNotes, extractWarnings, extractQuestions,
extractImports, extractArrowRecords, extractFieldTree, extractNamespaces.

CLI extract.ts becomes a thin re-export shim. 59 satsuma-core unit
tests; all 866 CLI tests continue to pass.
