+++ .tickets/sl-ikzl.md
@@ -1,6 +1,6 @@
 ---
 id: sl-ikzl
-status: open
+status: closed
 deps: [sl-kuos]
 links: []
 created: 2026-03-29T18:50:01Z
@@ -37,3 +37,9 @@ After this ticket, CLI's extract.ts becomes a thin re-export shim.
 
 1. satsuma-core/src/extract.ts exports all listed functions 2. extractFieldTree() is exported publicly 3. All functions import their helpers from satsuma-core's own cst-utils, classify, canonical-ref, meta-extract modules 4. satsuma-core builds without errors 5. New unit tests in satsuma-core/test/extract.test.js cover: schema extraction with nested record fields, mapping with each/flatten blocks, metric extraction, import extraction 6. CLI extract.ts becomes a re-export shim and existing CLI extract.test.js passes unchanged 7. Golden snapshot test (from sl-8pj3) still passes
 
+
+## Notes
+
+**2026-03-29T20:35:37Z**
+
+Cause: Extraction logic for schemas/fields/mappings/arrows/metrics/fragments/transforms/notes/warnings/questions/imports existed only in the CLI. Fix: Created satsuma-core/src/extract.ts with all extraction functions (importing from core's own cst-utils/classify/canonical-ref/meta-extract). CLI extract.ts is now a thin re-export shim. Added 59 satsuma-core unit tests. All 866 CLI tests pass.
